Record both types being compared.
The current code that compares types has protection against infinite
recursion. It works as follows.

For types which can introduce recursion (function types - hang on, can
they really? - and class/union types), a set of currently currently
being compared types is maintained and if a type is seen again, the
equality test automatically short-cuts to true.

The tests use of set containing both the left and right side of the
comparison and checks test for either. The difference between this and
the more correct insertion and checking of pairs may be nil or
non-existent but the storage requirements are likely to be almost
identical.

The bracketed push/pop behaviour means we can assert stronger
invariants on the code and only need to manipulate the top of a
stack. Using a vector over a set likely has a performance advantage as
the number of elements is typically very small.

- max size in test suite 10 pairs
- max size in kernel ABI 6 pairs
- typical max size 2-3 pairs

This commit switches from sets of single pointers to vectors of pairs
of pointers.

Signed-off-by: Giuliano Procida <gprocida@google.com>