Simplify type equality infinite recursion logic
Ignore this commit message, it's all been squashed. The intermediate
commits no longer passed the test suite, perhaps due to types where
the recursion ended up with a type on the left side of an equality
later appearing on the right.

------------------------------>CUT<------------------------------

> This is the 1st commit message:

Function types don't introduce new names so cannot cause self
references and so there is no need to track currently being compared
function types.

Typedefs do introduce new names but cannot refer directly to
themselves or be forward-declared. Enums introduce new names but
cannot refer to anything other than integral types. Therefore each
self reference must be via a class, struct or union type.

This commit removes the extra logic for function types.

Note that this short-cutting could actually be done at every level of
type comparison but it's unknown whether this would buy any
performance advantage.

	* src/abg-ir.cc (struct environment::priv): Drop member
	fn_types_being_compared_.
	(mark_as_being_compared): Drop function_type overload.
	(unmark_as_being_compared): Likewise.
	(comparison_started): Likewise.
	(equals): In the function_type overload, inline and remove the
	RETURN macro, bearing in mind that its argument was almost
	always false and the unmark_as_being_compared method is gone.
	(types_are_being_compared): Remove the code relating to
	function types being compared.

> This is the commit message #2:

abg-ir.cc: Adjust assertions around being-compared types

This commit adjust the assertions in the code that tracks in-progress
type comparisons. The mark/unmark operations should be strictly paired
and it should never be possible to pass a null pointer.

	* src/abg-ir.cc (unmark_as_being_compared): Adjust doc comment
	to reflect that the methods expect to find the type marked. In
	the overload that does the removal, assert that this happened.
	In the overload taking a plain pointer, remove the null
	pointer test. (comparison_started): In the class_or_union
        overload taking a plain pointer, remove the null pointer test.
        (mark_as_being_compared): Do something else. TBcompleted.

Signed-off-by: Giuliano Procida <gprocida@google.com>

> This is the commit message #3:

Mark/unmark/test comparisons part 1

> This is the commit message #4:

Record both types being compared.

The current code that compares types has protection against infinite
recursion. It works as follows.

For types which can introduce recursion (function types - hang on, can
they really? - and class/union types), a set of currently currently
being compared types is maintained and if a type is seen again, the
equality test automatically short-cuts to true.

The tests use of set containing both the left and right side of the
comparison and checks test for either. The difference between this and
the more correct insertion and checking of pairs may be nil or
non-existent but the storage requirements are likely to be almost
identical.

The bracketed push/pop behaviour means we can assert stronger
invariants on the code and only need to manipulate the top of a
stack. Using a vector over a set likely has a performance advantage as
the number of elements is typically very small.

- max size in test suite 10 pairs
- max size in kernel ABI 6 pairs
- typical max size 2-3 pairs

This commit switches from sets of single pointers to vectors of pairs
of pointers.

Signed-off-by: Giuliano Procida <gprocida@google.com>