Better DIE graph comparison.
The function compare_dies is performing a comparison between two
graphs, possibly subgraphs of the same graph. Any time we see a pair
of nodes being compared we can say "if we see these again, we can
assume equality succeeded". Note that there is no need to pop these
comparisons and indeed that can result in massive slowdowns.

This commit implements this strategy, replacing the following existing
logic:

- aggregates_being_compared a set of some subset of nodes being compared (LHS and RHS
  mixed together)
- logic to insert and erase nodes from this (not always paired up
  though)
- nodes representated as their string names

The comparisons are recorded as pairs of nodes as identified by
their DIE address.

Also, as a micro-optimisation, pointer identity of nodes is checked
first.

Signed-off-by: Giuliano Procida <gprocida@google.com>