Refresh ABI cross check test files
The test file test0-pr19026-libvtkIOSQL-6.1.so.1 is intended to be
used to check that diffing a binary entity against its ABI
representation results in an empty diff.  In this case, the ABI of the
library is also under revision control and so the test also functions
to a certain extent as check on whether the generated ABI is stable
between revisions of libabigail.

Recent changes to the DWARF reader have affected attributes such as
declared-inline and is-declaration-only. The result is that there is
now a non-empty diff between the library and the saved ABI, albeit all
"harmless" changes.

This commit refreshes the revision-controlled ABI, eliminating the
differences.

	* tests/data/test-diff-dwarf-abixml/test0-pr19026-libvtkIOSQL-6.1.so.1.abi:
	Refreshed ABI.
	* tests/data/test-diff-dwarf-abixml/test0-pr19026-libvtkIOSQL-6.1.so.1-report-0.txt:
	File now empty.

Signed-off-by: Giuliano Procida <gprocida@google.com>