abg-writer.cc: Track types by plain address
The XML writer tracks the state of referenced and emitted types so it
can emit the right types with no repetitions or omissions.

The sets and maps are keyed using type_hasher and deep_ptr_eq_functor
but these are known not to always agree when the types compared have
no canonical type.

The helper function type_is_emitted looks up types in a set but the
corresponding record_type_as_emitted stores the canonical pointer
where possible. Any discrepancy between structural and canonical
equality may result in types being treated as emitted when there are
not or vice versa.

Together, the result is that some types are unexpectedly duplicated or
omitted. Bugs due to inconsistent equality and hashing can be hard to
reproduce.

This change fixes the type_is_emitted discrepancy and switches the
maps and sets to be keyed by plain addresses. This does not fix the
underlying inconsistency but does prevent the XML writer from having
any role in decisions about type equality.

	* src/abg-writer.cc (type_hasher): Drop this helper class.
	(write_context): Change type_ptr_map, type_ptr_set_type and
	fn_type_ptr_set_type to use plain pointer equality and
	hashing. (type_is_emitted): Look up canonical type if it
	exists, to match record_type_as_emitted, get_id_for_type etc.
	* tests/data/test-read-dwarf/PR22122-libftdc.so.abi: Refresh.
	* tests/data/test-read-dwarf/test12-pr18844.so.abi: Refresh.
	* tests/data/test-read-dwarf/test9-pr18818-clang.so.abi:
	Refresh.
	* tests/data/test-read-write/test25.xml: Refresh.

Signed-off-by: Giuliano Procida <gprocida@google.com>