abg-writer.cc: Track types by plain address
The XML writer tracks the state of referenced and emitted types so it
can emit the right types with no repetitions or omissions.

The sets and maps are keyed using type_hasher and deep_ptr_eq_functor,
effectively resulting in structural equality checks in some cases.

The helper function type_is_emitted looks up types in a set but the
corresponding record_type_as_emitted stores the canonical pointer
where possible. Any discrepancy between structural and canonical
equality may result in types being treated as emitted when there are
not or vice versa, resulting in duplicate or omitted type ids.

Together, the result is that some types are unexpectedly conflated.
Also, depending on the toolchain and libraries, some types are not
emitted even when referred to elsewhere in the XML.

This change fixes the type_is_emitted discrepancy and switches the
maps and sets to be keyed by plain addresses.

	* src/abg-writer.cc (type_hasher): Drop this helper class.
	(write_context): Change type_ptr_map, type_ptr_set_type and
	fn_type_ptr_set_type to use plain pointer equality and
	hashing. (type_is_emitted): Look up canonical type if it
	exists, to match record_type_as_emitted, get_id_for_type etc.
	* tests/data/test-read-dwarf/PR22122-libftdc.so.abi: Refresh.
	* tests/data/test-read-dwarf/test12-pr18844.so.abi: Refresh.
	* tests/data/test-read-dwarf/test9-pr18818-clang.so.abi:
	Refresh.
	* tests/data/test-read-write/test25.xml: Refresh.

Signed-off-by: Giuliano Procida <gprocida@google.com>