CHECKME: Fix handling of anonymous union members
NOTE: This is a rough attempt based on similar code for structs. It
may have some unwanted side-effects (see the refreshed XML files).

This commit updates handling of unions with anonymous members in a
similar fashion to the way this is done for structs and classes.
It also refreshes a couple of test ABI files.

NOTE NOTE: This is what's left after Dodji's own fix.

	* src/abg-dwarf-reader.cc (add_or_update_union_type): Handle
	anonymous members in the same way as add_or_update_class_type.
	* src/abg-reader.cc (build_union_decl): Update a couple of
	comments to say "union" instead of "class". Check for existing
	members by value, not name.
	* tests/data/test-diff-dwarf-abixml/PR25409-librte_bus_dpaa.so.20.0.abi:
	Refreshed.
	* tests/data/test-read-dwarf/PR25007-sdhci.ko.abi: Refreshed.

Signed-off-by: Giuliano Procida <gprocida@google.com>