Simplify symbol alias representation
Symbol aliases are currently modelled as chain of "next" symbols
together with the notion of a main symbol. The chain is currently
closed to form a cycle which presents problems with iteration,
particularly if starting from an arbitrary position within the chain.

This commit makes the simple change that the chain starts with the
main symbol and terminates with a null pointer. This makes it easy to
iterate from or from just after the main symbols, without missing any
items and with a very simple loop termination condition that has no
risk of being confused by operator==(elf_symbol, elf_symbol).

The code revisits all loops over symbol list chains and make a variety
of (presumed) improvements, as follows.

The function does_alias was used in an M+N check for matching names.
This is replaced by a full MxN check implemented in elf_symbols_alias.

In many cases, the current and likely main symbol is now iterated
over. In some of these cases it was being treated specially, in others
it was omitted. In some cases if the symbol passed is not the main
symbol, then some aliases are now examined where they would have been
previously been skipped.

	* include/abg-ir.h (does_alias): Drop this.
	(compute_aliases_for_elf_symbol): Out argument alias_set now
	contains raw pointers for simplicity. Callers mustn't retain
	them.
	* src/abg-comp-filter.cc
	(function_name_changed_but_not_symbol): Just call
	elf_symbols_alias to work out whether the symbol is unchanged.
	* src/abg-corpus-priv.h
	(corpus::exported_decls_builder::priv::add_fn_to_id_fns_map):
	When associating a symbol and all its aliases with a function,
	loop over the entire alias chain starting from the main
	symbol, also use a for-loop instead of do-while and a goto.
	* src/abg-corpus.cc
	(corpus::priv::build_unreferenced_symbols_tables): Simplify
	symbol alias loops.
	* src/abg-dwarf-reader.cc
	(read_context::load_symbol_maps_from_symtab_section): Call
	elf_symbols_alias instead of defunct does_alias.
	* src/abg-irc.cc (elf_symbol): Update alias documentation.
	(elf_symbol::get_main_symbol): Change doc comments to
	emphasise that main symbols are the first in the chain.
	(elf_symbol::has_alias): Redefine to match new chain model.
	(elf_symbol::get_number_of_aliases): Likewise.
	(elf_symbol::add_alias): Likewise and note that the
	implementation can make arbitrary symbols into alias.
	(elf_symbol::get_alias_from_name): Simplify symbol alias loop.
	(elf_symbol::get_alias_which_equals): Likewise.
	(elf_symbol::get_aliases_id_string): Simplify.
	(elf_symbol::does_alias): Drop this.
	(elf_symbols_alias): Just do MxN comparison.
	(compute_aliases_for_elf_symbol): Adapt the code to the new
	model.
	* src/abg-suppression.cc
	(function_suppression::suppresses_function): Simplify symbol
	alias loops.
	* src/abg-writer.cc (write_elf_symbol_aliases): Simplify
	symbol alias loop.
	* tests/data/test-diff-filter/test30-pr18904-rvalueref-report0.txt:
	Add missing aliases to removed/added functions/variables not
	mentioned by debug info.
	* tests/data/test-diff-filter/test30-pr18904-rvalueref-report1.txt:
	Likewise.
	* tests/data/test-diff-filter/test30-pr18904-rvalueref-report2.txt:
	Likewise.
	* tests/data/test-diff-filter/test35-pr18754-no-added-syms-report-0.txt:
	Likewise.
	* tests/data/test-diff-filter/test35-pr18754-no-added-syms-report-1.txt:
	Likewise.
	* tests/data/test-diff-filter/test41-report-0.txt:
	Likewise.
	* tests/data/test-diff-pkg/tbb-4.1-9.20130314.fc22.x86_64--tbb-4.3-3.20141204.fc23.x86_64-report-0.txt:
	Likewise.
	* tests/data/test-diff-pkg/tbb-4.1-9.20130314.fc22.x86_64--tbb-4.3-3.20141204.fc23.x86_64-report-1.txt:
	Likewise.

Signed-off-by: Giuliano Procida <gprocida@google.com>