abidiff: better whitespace in symbol diff report
maybe_report_diff_for_symbol has a few issues:

1. There is diffuse responsibility for newline emission and indeed
would emit spurious blank lines before most of the sub-diffs it
reports.

2. Different sub-diff text and terminal commas would be emitted
according to whether or not there had been a previous sub-diff -
making the output harder to grep and post-process.

3. The function also returns a bool but this is never inspected.

Until the advent of CRC reporting, the function was barely exercised
by the test suite.

This change changes the function to return void, changes the function
stanzas to always emit newline-terminated lines and ensures the
wording and punctuation of each sub-diff does not vary. It also tweaks
(shortens) the wording used for CRC diffs.

	* src/abg-reporter-priv.cc (maybe_report_diff_for_symbol):
	Make return void. Simplify and fix new-line emission. Remove
	comma emission. Tweak CRC wording.
	* src/abg-reporter-priv.h (maybe_report_diff_for_symbol):
	Make return void.
	* tests/data/test-abidiff-exit/test-crc-report.txt: Tweak CRC
	wording.
	* tests/data/test-abidiff/test-crc-report.txt: Likewise.

Signed-off-by: Giuliano Procida <gprocida@google.com>