Add pass to drop empty XML elements
Certain elements in ABI XML are effectvely containers and can be
dropped if empty and their attributes don't carry ABI information.

- elf-variable-symbols: pure container
- elf-function-symbols: pure container
- namespace-decl: has a name
- abi-instr: compilation unit (path etc.)
- abi-corpus: binary object (architecture)
- abi-corpus-group: binary objects (architecture)

It could be argued that abi-corpus (or abi-corpus-group) should be
kept around to hold the architecture of an object or set of objects.
However, if a binary object has no symbols (say, if it is empty), it
hardly matters what the architecture is.

Note that:

- abidiff rejects XML files with an XML declaration at the top
- abidiff rejects completely empty files

Resolving the first would make the second moot. In the meantime, we
avoid dropping top-level elements.

	* scripts/abitidy.pl (drop_if_empty): New variable containing
	the tags of elements that can be dropped if empty.
	(drop_empty): New Function that removes empty elements, except
	top-level ones.

Signed-off-by: Giuliano Procida <gprocida@google.com>