Add pass to eliminate duplicate member-type fragments
One common kind of duplicate type in ABI XML occurs when the XNL
writer emits a type declaration with its surrounding scope but
that scope actually includes one or more user-defined types.

An additional complication is that a nested member type's access can
be misapplied to its containing type.

The types themselves are later emitted in full and with correct member
access specifiers.

	* scripts/abitidy.pl (sub_tree): New function to determine if
	one XML node is a sub-tree of another (in the sense that the
	former can be overlaid without changing the latter), ignoring
	possibly-incorrect access attributes.
	(eliminate_duplicate_types): New function that checks nodes
	with the same type id, determines if there is a maximal node
	and, if so, remmoves the other nodes.

Signed-off-by: Giuliano Procida <gprocida@google.com>